0.2.1 Ported Pipe.java to Pipe.scala
Once I've done the same to Truss.java & LightingStand.java, I will be able to make Mountable a trait
rather than having it be wedged into an inheritance tree that concerns itself with issues of parsing
and generating XML.

Why yes, it has been a long time since I've made a reasonable commit of just one change.
Putting this project back the way it should be is going to be a lot of work, but hopefully worthwhile.