0.2.14 Gave Venue the Populate trait so it can instantiate Proscenium and Pipe objects.
With this commit, I was only planning on populating a Venue with Pipes (so that
Parse would no longer be attempting to instantiate Pipes on PipeBases which
Event -> PipeBase -> Pipe had already created) but it turned out to break tests
because Venue -> Pipe associations were being processed before Parse had an
opportunity to instantiate any proscenium elements that were part of tests.

Having Venue populate a proscenium (if it exists) before any pipes took care of
that handily.

Note that Parse no longer looks for proscenium, pipebase, or pipe elements.
The following associations are supported:
  - Venue -> Proscenium
  - Venue -> Pipe
  - Event -> PipeBase
  - PipeBase -> Pipe

This means that pipebase elements cannot be situated as part of a venue. This
doesn't seem like it is likely to cause problems.