0.2.15 Show Luminaire detail data on mouseover.
This doesn't actually work. I am able to store the Luminaire data in a sibling element (using a separate namespace) and access that data via ECMAScript. I am able to embed an HTML table in the SVG document and set innerText content for its 'td' elements. But I am unable to display that embedded content once I hide it.

Also, my enthusiasm for this little side-project is blunted by the realization that I cannot run ECMAScript within any browser or SVG viewer I can find for Android. This code seems to do no harm, So I'm going to leave it in place for later reference while I move on to work that will actually be productive.