0.2.16 Added Flat.scala to represent a scenic flat.
This is so very much like a Wall (in particular the openings in a wall or flat) that I made ProtoWall to be a parent to both Flat and Wall. As part of this, I ported Wall to scala.

Created a Line class to simplify dealing with two points representing a line. Its two methods simply invoke Point methods, but having a Line helps in situations where repeated use of the same pair of Points might otherwise be obfuscated in the code.

Changed Proscenium.Locate(Point) to Proscenium.LocateIfActive(Point) because it was only ever used inside a guard to check Porscenium.Active(). Code in various classes code were able to be simplified.